Fix grammar of "command-line" where needed.
When used attributively, e.g. as an adjectival phrase, "command-line"
should be hyphenated.  See, for instance,
<http://www.chicagomanualofstyle.org/16/images/ch07_tab01.pdf>.

Also found and fixed one error in the reverse direction in tmac/s.tmac.

I did not bother to fix mistaken uses in change logs, NEWS files, and
similar, but did in source comments.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>