Put two spaces between sentences in GPL notice.
Fixes half of Savannah #51609 but applies it much more generally.

This is a wide-ranging change, but it makes these notices consistent
with modern copies of the GPLv3 text.  (Some of these files are not
GPLv3, but that's where the boilerplate comes from; GPLv2 and LGPLv2.x
used semicolons to separate these clauses instead.)

If the next resync with Autotools regresses some of these, well, that's
a bug in Autotools.

I checked with wc -L, and no file grew its longest line to 81 characters
due to this change.

contrib/hdtbl/hdmisc.tmac-u only:
Signed-off-by: Bjarni Ingi Gislason <bjarniig@rhi.hi.is>

Blame me for the rest.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>