test-groff.in: Recognize additional contrib tmacs.
"tmac" files in the directories "contrib/hdtbl", "contrib/mm", and
"contrib/pdfmark" are not included in the search path for "test-groff".
That is of cause necessary to test those macro files and files using
those.

Fixes Savannah #51598.

Signed-off-by: Bjarni Ingi Gislason <bjarniig@rhi.hi.is>

Break these hard-to-read declarations across lines so they're easier to
maintain.

Kill useless trailing whitespace.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>