groff_mmse(7): Stop marking up copyright notice.
This copyright notice is inside an .ig region and thus does not require
roff markup.

Revert my change in 117113104468fa6e0f5c63fa6d0febdbe88da826 which
changed the ASCII dash in the copyright year range to a character
escape.

The character escape for the copyright sign was present at git
import.

Thanks to Werner Lemberg for prompting me to check up on this.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>