Use gnulib script 'git-version-gen' to define groff's version
* bootstrap.conf: add 'git-version-gen' script.

* configure.ac: suppress 'GROFF_VERSION_STRING' and replace it by
the script.  If the current commit corresponds to a tag, then the
version is simply the tag name.  Otherwise the version has the
format: <tag>-<nb_commits_from_latest_tag>-<commit sha1>.  The
version is now stored in '.version'.

* REVISION: removed.

* VERSION: removed.

* Makefile.am: define 'MAJOR_VERSION', 'MINOR_VERSION' and
'REVISION'.

* src/libs/libgroff/libgroff.am: use 'MAJOR_VERSION',
'MINOR_VERSION' and 'REVISION', add dependency on '.version'.

* src/roff/troff/troff.am: ditto.

* contrib/mom/om.tmac-u (short_revision): new string that extracts
the first digit of the revision, to be used when checking groff
version.