pic2graph(1): Stop trying to outsmart .TP macro.
Remove gratuitous calculated indent setting.  Most man pages that do
this trick do it as a way to simulate 2-column output, or a 2-column
table, but in this case the table would have only one row so there's
no beautification to be done.

This trick causes 'can't break line' warnings if, for instance, the
./configure --prefix is longer than the man page author expects.  (This
is often the case.)

Thanks to Bjarni Ingli Gislason for catching this.

Fixes bug https://savannah.gnu.org/bugs/index.php?53715.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>