pic2graph: Refactor temp directory handling.
contrib/pic2graph/pic2graph.sh: Issue a diagnostic and give up if the
user's desired temporary directory does not exist.  Also document the
Bashism $RANDOM and work around its absence.

The work-around is crap entropy but slightly better than a null string
(/bin/dash's $RANDOM), and at this point in the logic we're already on a
system where the mktemp command is not installed, doesn't support the
flags we need, or doesn't work.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>