man pages: Add italic correction before '\c'.
The italic correction '\/' in the macros 'I', '[BR]I', and 'I[BR]' is
ignored if it is applied after a '\c' escape in an argument.

Signed-off-by: Bjarni Ingi Gislason <bjarniig@rhi.hi.is>

"\c works on the output level. Anything after this escape on the same
line is ignored except \R, which works as usual." (_The GNU Troff
Manual_, section 5.14, "Line Control")

Manually restore the italic correction where applicable.

And even where it's not; the next thing output after the italic
correction in the chem.1 and groff_man.1 cases is a space, which does
not receive an italic correction.  But it's harmless, seems better to
have it for pedagogical (or cargo-cult) reasons, and is more robust
against future changes that may put a glyph where the space had been.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>