groff_man(7): Update discussion of \c and \f.
Share the lesson learned from Savannah #53795.

Document interaction of .BI and .RI macros with italic corrections and
use of \c escape in final (if even-numbered) macro arguments.

Slightly clarify which argument is meant when discussing .TP; it's not
the optional one on the same line, but the one on the next line.

Recast software package names in a more prosaic way, instead of
a tar-file-ish one.

Get rid of a stray .sp request.

Admit that anyone who wants italic corrections is going to have to give
up some portability (presumably).

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>