groff_man(7): Relocate the Options section.
...to follow all the sections that are really just extensions of the
Description section, which is really just an introductory paragraph.

This commit is to prepare for conversion of those many aforementioned
sections to subsections, which are not presently used in this page.

The objective is to get the information that man page writers need to
know in front of them sooner rather than later, and particularly before
the use of groff options with the macro package, which is a pretty
esoteric use case.  Most people will feed the package register settings
via man(1), perhaps unknowingly, and if at all.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>