groff_man(7): Rearrange macro discussion.
Organize macro descriptions into more (sub)sections, adding "document
structure macros", "font face macros", and "paragraph macros".  Also add
a "deprecated features" subsequent, thus named so we can put more than
just macros into it, if necessary.

Retitle subsections covering command synopses and email addresses and
URLs to be shorter and fit in the table on an 80-column TTY.

Precede macro discussions with a table alphabetically listing all
non-deprecated macros and referring the reader to the subsection where
they are described.

Add paragraph mentioning deprecated macros not mentioned in the table.

Present macros in each subsection in a more pedagogically useful order.
For instance, the very first macro a person writing a man page from
scratch needs to know about is .TH.  (The previous ordering was not
strictly alphabetical anyway; we started with .EX, not the
strictly-first .AT, and not .B, the first non-deprecated macro.)

No alterations were made to relocated material for this commit.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>