groff_man(7): Remove italics from "man pages".
We don't need to _both_ set "man pages" in italics and quote it when
introducing it as jargon.  I favor quotation marks for it since it's
toward the slang end of the spectrum.  (Contrast with phrases like
"Hamming distance" or "NP-hard", which would be italicized upon their
first appearance in a formal text introducing the concepts.)

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>