tmac/groff_www.7.man: Do not use .URL, .MTO, and .FTP.
A manual page about a macro set must not use its own macro set,
but rather use man(7) or mdoc(7), or it is not portable.
It is trivial to fix that here because the groff_man(7) GNU
extensions offer identical functionality with .UR/.UE and .MT/.ME.

While here, also delete the duplicate FTP line from the macro list.

OK gbranden@