groff(7): Fix style nits.
When a tagged paragraph is cased as if the tag begins a sentence, it
should actually be a sentence.

Also give the claim of groff_diff(7) as an authoritative document its
own paragraph; this is a notable contrast to the several (incorrect)
claims we have in man pages characterizing the Texinfo manual as the
complete and authoritative source of all groff documentation.

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>