Further tweak shell script portability in the hdtbl fonts examples.
As gbranden@ rightly observed, the number of characters in the
character class :cntrl: is locale-dependent, POSIX says that the
repeat count in the second argument to tr can be omitted, and the
transliteration target will grow to fit the size of the source.
This actually works on OpenBSD, Debian Jessie, Solaris 11, and
Solaris 9, so it seems quite portable indeed.

People keep discussing all kinds of fancy changes on the mailing
list, but fail to address the question of which is the minimal
change required to get portable behaviour during the last minute
before release.  Thus, i'm just making a decision and using the
minimal tweak suggested by gbranden@.