Remove the unused variables PREFIXMAN5 and PREFIXMAN7.
They caused portability trouble because they resulted in
commands of the form:  "for f in ; do $something ; done"
Even though such empty loops are valid POSIX syntax,
they cause the default sh(1) on Solaris 9 to die.

Thanks to wl@ for insisting that the Solaris 9 failure
ought to be fixed; without that, the fact that these
variables are unused would not have been found.

OK bgarrigues@ gbranden@