src/roff/groff/tests/*: Remove redundant syntax.
"exit $?" at the end of a shell script means the same thing as just
plain "exit" which means the same thing as just falling off the end
of the script.  The exit status of the script is that of the last
command executed.

I can't find the exact spot in POSIX 2017 that mandates this, but none
of the following shells screw it up:
	bash 5.0
	dash 0.5.10.2
	ksh 93u+20120801
	mksh 57
	posh 0.13.2
	yash 2.48
	zsh 5.7.1