Repair .AT and .UC in the groff_man(7) macros.
* tmac/an-old.tmac:
Setting user-defined strings in a macro that will later be called
indirectly from page location traps is excessively complicated.
Besides, the implementation doesn't work: when the trap is finally
sprung, the defaults from the an-init macro clobber what the author
specified with .AT or .UC.
Instead, all that is needed is setting the strings for the header
before triggering the page break, such that they appear right
away, while setting the strings for the footer after the page
break, such that they don't appear on the previous page.

This bug was found by Jonathan Gray <jsg@openbsd.org>
while he looked at 4.xBSD manual pages.

Thanks to gbranden@ for finding a bug in first version of my patch
and for agreeing with the idea, see: https://savannah.gnu.org/bugs/?57416