Add regression test for Savannah #57416.
* tmac/tests/an-old_AT_and_UC_footer_saved_and_restored.sh: Add
  regression test for Ingo's change (previous commit).

* tmac/tmac.am: Add foregoing to tmac_TESTS.

* ChangeLog: Reflow Ingo's entry for that message to fit GNU ChangeLog
  style better (even though elements of that norm drive me crazy).

I couldn't manage to construct a case that would reproduce any breakage
in .UC, but .AT was easily broken, and Ingo's patch simplifies the
overall logic on top of fixing .AT so I have no objection to it.

If the footers should break in the future, we'll sure catch 'em.