doc/webpage.ms: Fix style problems.
Change of point size is too large, +-2 is usual.

Change a HYPHEN-MINUS (code 0x55, 2D) to a minus (\-), if in front of a
name for an option.

Put a '\&' directly after a period (.), that is not an end of sentence.

Use a straight quote (apostrophe, apostrophe quote, quotesingle, \[aq])
in examples (using a constant width font, for example Courier (C, CR).
See for example "www.cl.cam.ac.uk/~mgk25/ucs/quotes.html".

Use a hyphen to combine two words used as an attributive, see for example
"32-bit" in man-pages(7).

Use the name of units in text, not the symbol, except in tables,
calculations, and equations.

Use the unit 'p' (roff), not 'pt' (TeX).

Set the name of a man page in bold type, but the section in roman.

Make '~' (tilde) bigger (+2) or use "\[ti]", if it means "approximate",
as the character is too small in the output of troff.

Change <X>[\]-<Y> to <X>\[en]Y> if it is a range.

Add a comma after "e.g." and "i.e.", or use English words
(man-pages(7)).

Add the missing 'l' (landscape) in 'a4'.

[I have a few tweaks I'd like to make to some of the above, but I'll
place those in a separate commit. --GBR]

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>