src/roff/troff/input.cpp: Update comments.
Fix groff version reference in comment.  We did actually release 1.22.4.

Also bump copyright year; significant* non-cosmetic changes to this file
were made in 2019 and this year.  (I know that a script can be directed
to come around and bump the copyright year on this file even if, say,
only whitespace changes were made, or no changes at all, but I dislike
that practice.  In my view, "copyrightable work" is a legal concept
inaccessible to machines.)

* as in "measurable", not "important"