src/roff/troff/input.cpp: Rewind copyright year.
I reverted my recent changes to this file because Ingo hates them and
emailed me to tell me so.

For fun, throw the following at nroff:

.pl 1v
\s
\sA
\s+
\s-
\s++
\s--
\s(+
\s(-
\s(++
\s(--
\s1
\s1x
\s[+
\s[-
\s[++]
\s[--]
\s

Every one of these is malformed but the current code (in its
now-reverted state) misses a few, and sometimes characters get eaten
(instead of sent to output) because of the lack of error-checking in
read_size().  Ingo was right that my fixes were incomplete (IMO
read_size() needs a refactor), but my changes caught more problems (with
a higher volume of diagnostics, admittedly).

I'll work on something else.