Correct and clarify point size documentation.
The Texinfo manual has long documented the request '.ps 0' as restoring
the previous point size (just as '\s0' or '.ps' with no argument does),
but this is incorrect, since groff 1.02 or earlier (June 1991), the
request has not actually worked this way.  Instead, '.ps 0' sets the
point size to 1 basic unit (thought output drivers may clamp this to a
higher value).  This behavior is consistent with Hierloom troff and,
per Ingo Schwarze, Plan 9 troff.  (It is, however, not consistent with
neatroff.)

* doc/groff.texi (Changing Type Sizes):
  + Stop claiming that '.ps 0' works like '\s0'.
  + Note that the resulting (computed) point size, not the argument, is
    clamped.
  + Note that it is non-positive, not negative, point sizes that are
    clamped to 1u.
  + Add (forward) cross-reference to section where \n[.ps] is defined
    since it is mentioned here.
  + Move explanation of special handling of zero argument to \s escape
    description.

* man/groff.7.man: Explain .ps N independently of \s.  Note clamping
  behavior.