man/groff.7.man: Fix rendering nit.
The documentation of the '\s±(N', '\s±[N]', and "\s±'N'" syntax forms
was rendering the '±' in the bold used to indicate literal input instead
of the italics used for parametric input.  This is because those syntax
forms are unique to \s and the page-private macros that the page author
wrote in an apparent effort to "semantically lift" the page contents
were inadequate to express them.

My preference would be to ditch the page-private macros altogether but
that is a larger effort.

Note that in PostScript/PDF output, an italicized '±' does not appear in
italics.  I assume that is because it is drawn from the "symbol font" S,
which lacks styling.  This is not a regression; none of the existing '±'
glyphs in italic context render italicized or slanted either.  The
output now looks correct on terminals, however.