Catch nroff/groff version desync.
Note that this commit will likely lead to a failing test case (but
perhaps not, if your build environment lacks a system groff).

It's too easy for the nroff version to get desynced from the groff
version when we're running test cases, leading to spurious results.
Make it easier to see a discrepancy.

* src/roff/nroff/nroff.sh: Call groff with -v or --version when we are
  called that way.

* src/roff/nroff/tests/verbose_option_works.sh.in: Check the nroff
  version being tested against the groff version being wrapped.  This
  exposes a bug; the system groff rather than the build tree groff was
  being invoked.

Also delete an old comment in nroff.sh.  The file uses the Stephen
Bourne "brace style" now, so the issue is unlikely to recur.