Tweak Texinfo "Implementation Differences" again.
As an experiment, here's a word diff of changed sentences.  I didn't
touch much--mainly minor wording changes, consistent use of the Oxford
comma, and hauled a sentence over from the corresponding material in
groff_diff(7).

I also amended this commit to change "TTY" to "terminal", and remove the
phrase "for more information" after cross-references; I find it
redundant.
----

AT&T 'troff' also interprets '\*[' [-or-] {+and+} '\n[' as a reference
to a string or number [-register-] {+register, respectively,+} called
'['.

For example, [-the-] {+this+} code produces bold output in both
cases, but the text [-differs:-] {+differs.+}

GNU 'troff' does not allow the use of the escape sequences '\|', '\^',
'\&', '\{', '\}', '\<SP>', '\'', '\`', '\-', '\_', '\!', '\%', and '\c'
in names of strings, macros, diversions, number registers, [-fonts-]
{+fonts,+} or environments; AT&T 'troff' does.

Macros, [-diversions-] {+diversions,+} and strings are all, in fact, the
same type of object; they contain lists of input characters and glyph
nodes in any combination.  {+Special characters can be both: before
being added to the output, they act as input entities; afterwards, they
denote glyphs.+}

[-For-]  {+The following+} example,

     .di x
     \\\\
     .br
     .di
     [-.x-]
     {+.x,+}

To [-store, for some reason,-] {+store+} an escape sequence in a
diversion that is interpreted when the diversion is reread, either use
the traditional '\!' transparent output facility, or, if this is
unsuitable, the new '\?' escape sequence.

(2) Naturally, if you've changed the escape character, you need to
prefix the 'e' with whatever it is--and you'll {+likely+} get something
other than a backslash in the output.