NEWS: Make format regular.
Update fill column to 72 (for consistency with most other files in the
source tree) and reflow paragraphs accordingly.

Use two blank lines before version "sections" and only one before subsystem
"subsections".  This is in fact the convention in older parts of the file,
but it got lost relatively recently.

Update editor assistance comments.