man/groff_diff.7.man: Document \k( and \k[].
Document the '(' and '[]' forms of the \k escape.  The presence of '\k('
as a groff difference is not an error; CSTR #54 does not admit this
syntax form.

man/groff.7.man: Synchronize \k description wording with groff_diff(7).
Also eliminate '\~\c' antipattern (when not needed in an input trap) in
favor of better macro usage.