Update documentation of .pm request.
groff(7) was just plain wrong, documenting classical behavior instead
(and omitting mention of strings and diversions).  Correct it.

groff_diff(7) now documents the difference between AT&T troff and
groff.

The Texinfo manual was accurate but did not mention the implementation
difference in the relevant section.  I moved it and recast some wording.