preconv(1): Reorganize.
Move compatibility mode save below legal terms notice, since we don't
need compatibility mode in that comment block or the apropos/mkwhatis
line before it.

Coalesce "Usage" section into "Description"; we don't need both, and
"Usage" is nonstandard per Kerrisk's man-pages(7) (which is not
normative for us, but useful and suggestive).

Except for replacement of the "Usage" section heading with a paragraph
macro call, this is a relocation-only commit.