doc/groff.texi: Use more appropriate markup.
The Texinfo manual says that @code{} is for "entire syntactic tokens"
(implicitly, not more), while @samp{} is for "literal example[s] or
'sample[s]' of a sequence of characters".

Convert several examples of multi-token @code{} samples to @samp{}.
Because samples are surrounded by quotation marks, the ambiguity
potentially produced by having them interrupted by a line break is
reduced, so also kemove usage of @w{} where applicable.

The construction "the @samp{groff -ms} macros" might now be more
awkward; but in my opinion the historical usage ("the -ms macros") was
awkward already.  We can come up with something more mellifluous if this
sticks in the craw.