Recast descriptions of .aln and .rr requests.
"The new name and the old name are exactly equivalent." was not helpful
to me in understanding the alias mechanism.  Moreover, aliases in groff
do not work like lexical substitutions (which are a significant theme in
the rest of the language).

Parallelize descriptions between Texinfo manual and groff_diff(7), and
use abbreviated description in groff(7).  Use same metasyntactic
argument names in all three places.  The pair "new" and "old" are not
perfect, but better than "xx" and "yy" (follwing CSTR #54) which
communicate nothing; "ident1" and "ident2" (same); or "alias" and
"register" which could imply a kind of second-order or subordinate
status for the alias.  I think this is what "The new name and the old
name are exactly equivalent." was trying to get across, but there is
obviously a sense in which they aren't--strcmp(), for example (although
".aln a a" on an existing number register 'a' does work as a no-op).

Update description of .rr request to note that, technically, it deletes
names rather than the objects themselves.  Parallelize metasyntactic
argument name between Texinfo manual and groff(7).