NEWS: Fix confusing sentence.
The statement was puzzling as worded, because it seemed to refer to two
different things as "AT&T troff".  This is in fact accurate and sound
(as far as I know), but confusing to the modern reader unaccustomed to
the unversioned rolling development practices at Bell Labs.  It was
famously observed that the "editions" of Research Unix came from the
versions assigned to printings of the Unix _manuals_, and that whatever
_code_ comprised the system was whatever happened to be live on the disk
at that time.