preconv(1): Fix coding tag grammar description.
The semicolon after the value portion of a coding tag declaration is
optional, not mandatory.  Thanks to Dave Kemper for the report.

Also fix a typo.

And add \& after the final dot in an ellipsis, because when I last
updated this page I had forgotten that ']' had its 5th cflags bit set
(i.e., 32; i.e., it's transparent for end-of-sentence recognition).  It
doesn't technically matter here because the context is an .EX/.EE block,
and thefore adjustment is off, but it is IMO a salutary example to set.