groff_char(7): Eliminate raw request usage.
Discard the rest of the low-level request-flinging.  It turns out that
it was invalidated by commit 1a44514ecfd5d1c3e46dfce91b531fa0876657a4
(August 2012).  The DVI device font fallback (conditioned on the ECFONTS
register) from CR to R only made sense if anything in the page tried to
access font CR, which it didn't after that commit.

Similarly, the string for the apostrophe (neutral single quote) was
defined for the benefit of some involved logic prior to that commit that
attempted to get the appearance of one in the Computer Modern fonts the
DVI device uses.  This was no longer functional after its removal.
While it's likely this is a regression, the regression happened in 2012.
(I'd be less equivocal, but I have not learned how to build groff before
the build system got redesigned around bootstrap, gnulib, and Automake.)

Moreover, my own (short-lived) comments about the string definitions
were true but beside the point--the \*(aq string was used in only one
place; plenty of other \[aq] escapes are present in the rest of the
file.  Beyond that, the page crawls with groff-style escapes, so any
troff that doesn't support \(aq is certainly going to choke hard on lots
of other things in the page (or simply display nonsense).

Therefore:

* (page prologue):
  + Discard the aq string definition and its fallback on non-groffs.
  + Replace its one dereference with an \[aq] escape.
  + Remove special-font fallback for unused font.

* (Description): Wordsmith some infelicitous language.

* (Reference/Named glyphs/Brackets): Replace .EX/.EE-replicating
  requests with .EX and .EE.

* (tables generally): Stop bracketing .TS/.TE requests with .na/.ad.
  They do nothing on grotty(1) and grops(1).