doc/groff.texi: Relocate some string material.
This is a relocation-only commit (almost) in preparation for editorial
work.

The exception involves movement of the .T string description from the
node "Built-in Registers" to the node "Strings".  The former was in a
table, so adjust whitespace, drop the @item macro and wrap it in (local)
Defstr macros at its destination to prevent parsing problems.