doc/groff.texi (Strings): Relocate material.
Defining .ds1 in the middle of a large volume of text about features of
strings in general interrupts the flow of the narrative in a bad way.
Push .ds1 to the end of the .ds/.ds1/\* definition material; it is
relatively brief and none of the foregoing discussion is inapplicable.