doc/groff.texi, groff_diff(7): Use term "request".
...in favor of "command".  Except for an introductory portion of
groff(7), our documentation is pretty good about reserving the term
"command" to Unix commands, instructions in the bespoke formats of
device and font description files and the intermediate output language,
or drawing commands, the last being the sublanguage of the \D escape.

I did not fix groff(7) because it needs a more considered rewrite to
overcome this problem.