groff(7): Fix string and double-quote error.
man/groff.7.man (Control Characters): Incorrect claims were made.
Double quotes can indeed be used to enclose arguments to string
interpolations.

	.ds foo foo \\$1 baz \\$2 and then \\$3\"
	This is \*[foo bar "qux" "]"].
		This is foo bar baz qux and then ].

The statement about a leading " in a string definition (or appendment
with .as) was stated too generally.  The leading quote is not necessary
to include leading tab characters, which can be input as-is, even in
compatibility mode.

	(Imagine that multiple adjacent spaces are tabs below.)

	.ds a a
	.ds b	b
	My string is x\*ax.
	.br
	Another string is y\*by.
		My string is xax.
		Another string is y	by.

Also, use groff-style escape to document special character \[dq].