groff_man(7): Rename relative "indent" to "inset".
Inspired by a comment of Michael Kerrisk on the linux-man mailing list.

Due to the horrible confusion between "identation" and the "left margin"
in this man page historically, discuss the action of the .RS and .RE
macros in terms of a "relative inset", which is equally accurate in my
view.  It is often a bad idea to use the same term to refer to two
slightly different things.

Also expand the explanation of the IN register (-rIN option to *roff) to
note that it affects not just .LP/.PP/.P, but also .IP, .RS, .TP, .TQ,
and the deprecated .HP.