doc/groff.texi: Fix quotes in \' and \` escapes.
That is, get correct Unicode basic Latin (ASCII) glyphs to render.

Do the same for all escapes taking an arbitrary delimiter, for which the
apostrophe is the idiomatic choice.

Also mark up the apostrophes in the string-comparison conditional this
way.

Also convert the .tl request from using @t{} for its apostrophes (which
still produce directional quotes, just in a monospaced font) to @code{}.

To work, the above have to be seen by Texinfo when
	@codequotebacktick on
	@codequoteundirected on
are in effect.